add data to profile page
